+++ src/Microsoft.Health.CosmosDb/Microsoft.Health.CosmosDb.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
-  
+
   <ItemGroup>
     <PackageReference Include="Ensure.That" Version="8.1.2" />
     <PackageReference Include="Microsoft.Azure.DocumentDB.Core" Version="2.9.2" />
@@ -11,10 +11,10 @@
     <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks" Version="3.1.3" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="3.1.3" />
     <PackageReference Include="Microsoft.Extensions.Options" Version="3.1.3" />
-    <PackageReference Include="Microsoft.Health.Abstractions" Version="1.0.0-master-20200413-1" />
-    <PackageReference Include="Microsoft.Health.Core" Version="1.0.0-master-20200413-1" />
-    <PackageReference Include="Microsoft.Health.Extensions.DependencyInjection" Version="1.0.0-master-20200413-1" />
+    <PackageReference Include="Microsoft.Health.Abstractions" Version="1.0.0-master-20200416-1" />
+    <PackageReference Include="Microsoft.Health.Core" Version="1.0.0-master-20200416-1" />
+    <PackageReference Include="Microsoft.Health.Extensions.DependencyInjection" Version="1.0.0-master-20200416-1" />
     <PackageReference Include="Polly" Version="7.2.0" />
   </ItemGroup>
-  
+
 </Project>
