--- src/Microsoft.Health.Fhir.Core/Features/Persistence/ITransactionHandler.cs
@@ -1,14 +1,0 @@
-﻿// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-using System;
-
-namespace Microsoft.Health.Fhir.Core.Features.Persistence
-{
-    public interface ITransactionHandler : IDisposable
-    {
-        ITransactionScope BeginTransaction();
-    }
-}
